+++ packages/web3-rpc-methods/test/unit/eth_rpc_methods/fixtures/get_balance.ts
@@ -32,5 +32,7 @@ export const testData: TestData[] = [
 	['blockNumber = BlockTags.LATEST', [address, BlockTags.LATEST]],
 	['blockNumber = BlockTags.EARLIEST', [address, BlockTags.EARLIEST]],
 	['blockNumber = BlockTags.PENDING', [address, BlockTags.PENDING]],
+	['blockNumber = BlockTags.SAFE', [address, BlockTags.SAFE]],
+	['blockNumber = BlockTags.FINALIZED', [address, BlockTags.FINALIZED]],
 	['blockNumber = "0x4b7"', [address, '0x4b7']],
 ];