+++ packages/web3-types/src/eth_types.ts
@@ -45,8 +45,11 @@ export enum BlockTags {
 	EARLIEST = 'earliest',
 	LATEST = 'latest',
 	PENDING = 'pending',
+	SAFE = 'safe',
+	FINALIZED = 'finalized',
 }
-export type BlockTag = 'earliest' | 'latest' | 'pending';
+export type BlockTag = `${BlockTags}`;
+
 export type BlockNumberOrTag = Numbers | BlockTag;
 
 export interface Proof {