+++ packages/web3/src/types.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { Address, Bytes, ContractAbi, Transaction } from 'web3-types';
+import { Bytes, Transaction } from 'web3-types';
 import Eth from 'web3-eth';
 import {
 	decodeLog,
@@ -36,13 +36,10 @@ import {
 	Wallet,
 	Web3Account,
 } from 'web3-eth-accounts';
-import Contract, { ContractInitOptions } from 'web3-eth-contract';
+import { Contract } from 'web3-eth-contract';
 import { ENS } from 'web3-eth-ens';
 import Net from 'web3-net';
 import { Iban } from 'web3-eth-iban';
-
-export type Web3ContractConstructor<Abi extends ContractAbi> = Omit<typeof Contract, 'new'> &
-	(new (jsonInterface: Abi, address?: Address, options?: ContractInitOptions) => Contract<Abi>);
 
 /**
  * The Ethereum interface for main web3 object. It provides extra methods in addition to `web3-eth` interface.
@@ -60,8 +57,7 @@ export interface Web3EthInterface extends Eth {
 	 * web3.eth.Contract.setProvider(myProvider)
 	 * ```
 	 */
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	Contract: Web3ContractConstructor<any>;
+	Contract: typeof Contract;
 	Iban: typeof Iban;
 	net: Net;
 	ens: ENS;