+++ packages/web3-eth-contract/CHANGELOG.md
@@ -227,3 +227,4 @@ const transactionHash = receipt.transactionHash;
 ### Added
 
 -   Added functionality of `createAccessList` for contracts ( #5780 )
+-   Added support of `safe` and `finalized` block tags (#5823)